requirements.txt
PySocks
rich
requests
